feat: emove the commented or not used codes
